requirements/common.txt
Django==1.4.1
Fabric==1.4.3
Pillow==2.3.0
South==0.7.6
django-crispy-forms==1.2.0
django-grappelli==2.4.0
psycopg2==2.4.5
yolk==0.4.3
django-modeltranslation==0.3.3
django-dotenv==1.2
pygeocoder==1.2.2
